[Saad] Refactor. Commercial Data Processing program completed

+++ Commercial_Data_Processing/CreateAccount.cs
@@ -1,12 +1,13 @@
 ﻿//-----------------------------------------------------------------------
 // <copyright file="CreateAccount.cs" company="BridgeLabz">
-//     Company copyright tag.
+//     Copyright © 2020 Company="BridgeLabz"
 // </copyright>
 // <creator name="Saad Shamim"/>
 //-----------------------------------------------------------------------
 
 namespace Object_Oriented
 {
+    using System;
     using System.IO;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
@@ -23,13 +24,21 @@ namespace Object_Oriented
         public static void Create(string fileName)
         {
             string path = @"C:\Users\ye10398\source\repos\saadshamim95\Programming\Object Oriented\Commercial_Data_Processing\" + fileName + ".json";
-            string print = "{\n" +
+            if (File.Exists(path))
+            {
+                Console.WriteLine("Account with " + fileName + " already exist!!!");
+            }
+            else
+            {
+                string print = "{\n" +
                 "'" + fileName + "':[\n" +
                 "]\n" +
                 "}";
-            var record = JObject.Parse(print);
-            string jsonResult = JsonConvert.SerializeObject(record, Formatting.Indented);
-            File.WriteAllText(path, jsonResult);
+                var record = JObject.Parse(print);
+                string jsonResult = JsonConvert.SerializeObject(record, Formatting.Indented);
+                File.WriteAllText(path, jsonResult);
+                Console.WriteLine("Account Successfully Created!!!");
+            }
         }
     }
 }
